feat: deterministic game sessions via --seed with per-VM seeded Lua and RNG
- add GameSessionConfig with 64-bit session seed passed into Room and RoomRuntime
- support --seed <n> in client and headless server entry points, seeding global RNG, QHash, and per-session seeds
- derive per-session seeds from a base (seed + session index); log #GameSeed on room start
- add lua_newstate_seeded with deterministic hash seed and seeded CreateLuaState overloads
- route Lua math.random/randomseed through the game RNG so gameplay VM randomness is reproducible
- widen GameRng/qsanSeedRandom/room-runtime/ai-runtime seed APIs to quint64
- drop time()-based re-seeding in boss-mode, godLottery, and addRobotCommand

+++ src/server/room-runtime.h
@@ -53,7 +53,7 @@ public:
     quint64 nextDecisionId() { return ++m_nextDecisionId; }
     quint64 stateRevision() const { return m_stateRevision; }
     void advanceStateRevision(StateMutation mutation);
-    void seedRandom(quint32 seed);
+    void seedRandom(quint64 seed);
 
     void addPackage(Package *package);
     void setPackage(Package *package);
